animation implimentation
Got the spunch to move down the rows. I do not know if it accurately reflects where the spikes and player are in the 3d off screen stage. I do not know how to tie the two. I used the same logic as the 3d version, tying 4 spikes to a parent object and moving that parent object.